+++ CHANGELOG.md
@@ -15,6 +15,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Fixed the problem that in `summarizeFasta` output the order of variant sources in the same group is not consistent across runs. #428
 
+- Argument `--ignore-missing-source` added to `summarizeFasta` so sources not present in any GVF can be ignored without raising any error. #436
+
 - In `filterFasta`, when filter with expression table, changed to filter out peptides smaller than, instead of smaller or equal to, the value of `--quant-cutoff`.
 
 - Fixed the issue that in `splitFasta`, variant sources are not grouped as they are specified by `--group-source` #439